Add safety measure to database connections
We should never allow the test database name to match the production
database name.  By adding this assertion we ensure that there is almost
no chance that a data-destructive test be invoked on real data.  This
overreaches slightly (since database name is just one aspect that could
make a database unique) but the benefits of being even safer outweigh
the minor potential inconvenience of having to rename a test database.

Issue #317