Remove test:ci command
When we set up CI we added a `test:ci` command.  This was because our
normal test command was set to use --verbose, and we didn't need all
that noise in our CI.

The issue is that `test` is actually a special command, which will
invoke `pretest` automatically.  We want our CI to benefit from these
hooks.

The other problem, by the way, is that --verbose is super noisy
everywhere including on dev and so we actually shouldn't be forcing it
to be flagged every time we want to run `yarn test` anyway.

This removes `test:ci` and makes `test` less noisy.

Issue #317