Don’t check scrape time if beyond scrape horizon
Previously, before scraping a CNN transcript URL, we performed a
preflight check in this order:

1. Check the database: Have we ever scraped this URL before?
2. Extract the transcript date: Is it beyond our configured date
   horizon?

With that order, we _always_ hit the database, even when we’re going to
reject the scrape in the very next check.

Since the order is not technically meaningful, we can reverse it to
save ourselves many unnecessary database lookups.

This commit does just that.

Affects #277

Signed-off-by: Justin Reese <justin@justinreese.com>