Retry unsuccessful CNN scrapes
Before this change, we wouldn’t re-scrape a CNN URL that we’d already
seen before, even if the scrape was unsuccessful. This protects us from
repeatedly scraping broken links in perpetuity, but also means if a URL
was temporarily broken, we don’t retry it later.

With the `SCRAPE_DAY_HORIZON` guard protecting us from scraping
permanently-broken URLs in perpetuity, we’re going to go ahead and
retry scraping broken URLs.

Affects #337

Signed-off-by: Justin Reese <justin@justinreese.com>