Update test script to clean DB
We run migrations before tests, but we do not actually reset the test
database. This was actually an oversight, but became an issue as I was
testing migrations locally.

We now roll back all migrations before running test migrations.

Although this could cause certain problems, and is not efficient, it is
the best that Sequelize has to offer at the moment.  I've created an
issue requesting support for a more straightforward `db:reset` CLI call:

https://github.com/sequelize/cli/issues/881

And also opened a Stack Overflow question in case the community has
better thoughts:

https://stackoverflow.com/questions/61254567/resetting-a-sequelize-database