Add twitter list scrape initiation queue
In order to actually invoke the twitter list scrapers we need the scrape
initiation queue.  This is a similar pattern to twitter account
scraping, where we have a list of items that need to be scraped and the
initiation queue is responsible for creating a new job for each item on
the list at the scheduled time.

This commit includes some intentional technical debt.  Speficically, we
are using constants in a bad way.  The google document ID associated
with a twitter list should really be stored in a database.  However, the
newsletters themselves and their associations with twitter lists should
be configured via database in the first place.

Rather than create a new table that will be deleted once we move to
dynamic newsletters, we're going to create this debt and remove it
cleanly when adding that feature.  Part of the logic here is that we're
planning to add that feature next.

Issue #229