Implement dynamic newsletters
Newsletters used to be driven by constants and static templates. This
meant that adding a newsletter would involve:

1. Adding new twitter list constants.
2. Adding a new newletter handlebars template.
3. Adding a new newsletter worker.
4. Adding a new newsletter delivery queue.

This was acceptable when we only had two, but now it's clear we want to
be able to set up newsletters for a wider variety of locations.

This commit changes the way newsletters are generated so that the
variable portion (e.g. twitter lists, boilerplate language, email
subject) is driven by the database.  This allows us to use a single
queue, worker, and template.

This commit includes some sample newsletter seeds which can be
optionally installed.

Issue #360